MINOR; Small optimizations in `ReplicaManager#becomeLeaderOrFollower` (#11225)
This patch refactors `ReplicaManager#becomeLeaderOrFollower` to avoid having to re-iterate over all the partitions to determine which ones should become leaders and which ones should become followers.

The patch also refactors how partitions are marked as offline when the log can't be created. Before the patch, we were iterating over all the partitions in the request or in the delta to mark them as offline is the log was not present. Now, we mark them as failed directly if the log can not be created.

Reviewers: Luke Chen <showuon@gmail.com>, Jason Gustafson <jason@confluent.io>