KAFKA-13079: Forgotten Topics in Fetch Requests may incorrectly use topic IDs (#11104)
The FetchSessionHandler had a small bug in the session build method where we did not consider building a session where no partitions were added and the session previously did not use topic IDs. (ie, it was relying on at least one partition being added to signify whether topic IDs were present)

Due to this, we could send forgotten partitions with the zero UUID. This would always result in an exception and closed session.

This patch fixes the logic to check that any forgotten partitions have topic IDs. There is also a test added for the empty session situation when topic IDs are used and when topic names are used.

Reviewers: David Jacot <djacot@confluent.io>, Jun Rao <junrao@gmail.com>